requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.12.0

# Vector database
chromadb>=0.4.22

# Web framework
streamlit>=1.31.0

# HTTP requests for Gong API
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Text processing
tiktoken>=0.5.2

# Optional: For Pinecone instead of Chroma
# pinecone-client>=3.0.0
# langchain-pinecone>=0.0.3
